Fix user groups not being dynamically setup
The id of the authenticated user was not being returned by the authenticate function,
which caused the user not to be loaded properly later on in PluggableAuthLogin.php
execute().

This is fixed in master by commit 44041721513f3108e56d3d8d607bfca455e75264,
but that one can't be cleanly ported back.